+++ Smi/src/SmiScenarioTree.hpp
@@ -225,7 +225,7 @@ public:
 	/** get number of scenarios */
 	int getNumScenarios()
 	{
-		return leaf_.size();
+		return (int) leaf_.size();
 	}
